+++ components/measurement/MeasurementContainer.js
@@ -11,6 +11,8 @@ import HttpHeaderFieldManipulationDetails from './nettests/HTTPHeaderFieldManipu
 import HttpInvalidRequestLine from './nettests/HTTPInvalidRequestLine'
 
 import VanillaTorDetails from './nettests/VanillaTor'
+import PsiphonDetails from './nettests/Psiphon'
+import TorDetails from './nettests/Tor'
 
 import DefaultTestDetails from './nettests/Default'
 
@@ -23,7 +25,9 @@ const mapTestDetails = {
   telegram: TelegramDetails,
   http_header_field_manipulation: HttpHeaderFieldManipulationDetails,
   http_invalid_request_line: HttpInvalidRequestLine,
-  vanilla_tor: VanillaTorDetails
+  vanilla_tor: VanillaTorDetails,
+  psiphon: PsiphonDetails,
+  tor: TorDetails
 }
 
 // FIXME to have header and stuff